Move headers to diff_kit
