+++ perun/view_diff/flamegraph/run.py
@@ -95,59 +95,6 @@ def get_uids(profile: Profile) -> set[str]:
     return set(df["uid"].unique())
 
 
-def generate_headers(
-    lhs_profile: Profile, rhs_profile: Profile
-) -> tuple[list[tuple[str, Any, str]], list[tuple[str, Any, str]]]:
-    """Generates headers for lhs and rhs profile
-
-    :param lhs_profile: profile for baseline
-    :param rhs_profile: profile for target
-    :return: pair of headers for lhs (baseline) and rhs (target)
-    """
-    lhs_header = generate_header(lhs_profile)
-    rhs_header = generate_header(rhs_profile)
-    return lhs_header, rhs_header
-
-
-def generate_header(profile: Profile) -> list[tuple[str, Any, str]]:
-    """Generates header for given profile
-
-    :param profile: profile for which we are generating the header
-    :return: list of tuples (key and value)
-    """
-    command = " ".join([profile["header"]["cmd"], profile["header"]["workload"]]).strip()
-    machine_info = profile.get("machine", {})
-    return [
-        (
-            "origin",
-            profile.get("origin", "?"),
-            "The version control version, for which the profile was measured.",
-        ),
-        ("command", command, "The workload / command, for which the profile was measured."),
-        (
-            "collector command",
-            log.collector_to_command(profile.get("collector_info", {})),
-            "The collector / profiler, which collected the data.",
-        ),
-        (
-            "kernel",
-            machine_info.get("release", "?"),
-            "The underlying kernel version, where the results were measured.",
-        ),
-        ("host", machine_info["host"], "The hostname, where the results were measured."),
-        (
-            "cpu (total)",
-            machine_info.get("cpu", {"total": "?"}).get("total", "?"),
-            "The total number (physical and virtual) of CPUs available on the host.",
-        ),
-        (
-            "memory (total)",
-            machine_info.get("memory", {"total_ram": "?"}).get("total_ram", "?"),
-            "The total number of RAM available on the host.",
-        ),
-    ]
-
-
 def generate_flamegraphs(
     lhs_profile: Profile,
     rhs_profile: Profile,
@@ -229,7 +176,7 @@ def generate_flamegraph_difference(
         kwargs.get("height", DEFAULT_HEIGHT),
         kwargs.get("width", DEFAULT_WIDTH),
     )
-    lhs_header, rhs_header = generate_headers(lhs_profile, rhs_profile)
+    lhs_header, rhs_header = diff_kit.generate_headers(lhs_profile, rhs_profile)
 
     env = jinja2.Environment(loader=jinja2.PackageLoader("perun", "templates"))
     template = env.get_template("diff_view_flamegraph.html.jinja2")
