+++ perun/view_diff/report/run.py
@@ -16,7 +16,6 @@ from perun.utils import log
 from perun.utils.common import diff_kit, traces_kit
 from perun.profile.factory import Profile
 from perun.profile import convert
-from perun.view_diff.flamegraph import run as flamegraph_run
 from perun.view_diff.table import run as table_run
 
 
@@ -211,7 +210,7 @@ def generate_html_report(lhs_profile: Profile, rhs_profile: Profile, **kwargs: A
     ]
 
     env = jinja2.Environment(loader=jinja2.PackageLoader("perun", "templates"))
-    lhs_header, rhs_header = flamegraph_run.generate_headers(lhs_profile, rhs_profile)
+    lhs_header, rhs_header = diff_kit.generate_headers(lhs_profile, rhs_profile)
     template = env.get_template("diff_view_report.html.jinja2")
     content = template.render(
         lhs_tag="Baseline (base)",
